create kv_connector mixin for model runners
